Clear feedbackMap after receiving new feedback
Usually, the old feedback would simply be overwritten by the new feedback. If a robot is turned off or removed, however, no new feedback is given and the old feedback would be kept. This is undesirable as we would then receive outdated feedback for robots that are no longer in play. To prevent this, the map is now cleared before processing the new feedback.